requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
et-xmlfile==1.1.0
numpy==1.23.4
openpyxl==3.0.10
pandas==1.5.0
pykerberos==1.2.4
pymongo==3.12.3
python-dateutil==2.8.2
python-snappy==0.6.1
pytz==2022.4
six==1.16.0
sqlparse==0.2.4
tzdata==2022.6
